+++ psp/clients/uk_pv/scripts/simplify_data.py
@@ -9,7 +9,7 @@ import pathlib
 import numpy as np
 import pandas as pd
 
-from psp.data.uk_pv import C, filter_rows, trim_pv
+from psp.clients.uk_pv.data import C, filter_rows, trim_pv
 
 
 def parse_args():
